Move Maps to common
