+++ stateful-functions-flink/stateful-functions-flink-core/src/main/java/com/ververica/statefun/flink/core/translation/IngressToSourceFunctionTranslator.java
@@ -17,6 +17,7 @@
 package com.ververica.statefun.flink.core.translation;
 
 import com.ververica.statefun.flink.core.StatefulFunctionsUniverse;
+import com.ververica.statefun.flink.core.common.Maps;
 import com.ververica.statefun.flink.io.spi.SourceProvider;
 import com.ververica.statefun.sdk.io.IngressIdentifier;
 import com.ververica.statefun.sdk.io.IngressSpec;
