+++ stateful-functions-flink/stateful-functions-flink-core/src/main/java/com/ververica/statefun/flink/core/translation/RouterTranslator.java
@@ -18,6 +18,7 @@ package com.ververica.statefun.flink.core.translation;
 
 import com.ververica.statefun.flink.core.StatefulFunctionsJobConstants;
 import com.ververica.statefun.flink.core.StatefulFunctionsUniverse;
+import com.ververica.statefun.flink.core.common.Maps;
 import com.ververica.statefun.flink.core.message.Message;
 import com.ververica.statefun.sdk.io.IngressIdentifier;
 import com.ververica.statefun.sdk.io.IngressSpec;
